+++ tests/jobrunner/test_validators.py
@@ -2,165 +2,184 @@ import json
 from copy import deepcopy
 
 import bmds
-import jsonschema
 import pytest
 from django.core.exceptions import ValidationError
 
 from bmds_server.jobrunner import validators
 
 
-def test_invalid_json():
-    # invalid JSON
-    with pytest.raises(ValidationError):
-        validators.validate_input("{")
+class TestInputValidation:
+    def test_invalid_json(self):
+        # invalid JSON
+        with pytest.raises(ValidationError):
+            validators.validate_input("{")
+
+    def test_bmds2_model_override(self, complete_continuous):
+        # complete check
+        c_models = [{"name": "Exponential-M2"}, {"name": "Exponential-M3"}]
+        d_models = [{"name": "Logistic"}, {"name": "LogLogistic"}]
+        data = deepcopy(complete_continuous)
+
+        data["models"] = c_models
+        assert validators.validate_input(json.dumps(data)) is None
+
+        data = deepcopy(complete_continuous)
+        data["models"] = d_models
+        with pytest.raises(ValidationError) as err:
+            validators.validate_input(json.dumps(data))
+        assert "is not one of" in str(err)
+
+        # should fail w/ bmds3
+        data = deepcopy(complete_continuous)
+        data.update(bmds_version=bmds.constants.BMDS312, models=c_models)
+        with pytest.raises(ValidationError) as err:
+            validators.validate_input(json.dumps(data))
+
+    def test_bmds2_bmr_override(self, complete_continuous):
+        # Check models can be specified
+        c_bmr = {"type": "Std. Dev.", "value": 1.0}
+        d_bmr = {"type": "Added", "value": 0.1}
+
+        # complete check
+        data = deepcopy(complete_continuous)
+        data["bmr"] = c_bmr
+        assert validators.validate_input(json.dumps(data)) is None
+
+        # wrong bmr specification
+        data = deepcopy(complete_continuous)
+        data["bmr"] = d_bmr
+        with pytest.raises(ValidationError) as err:
+            validators.validate_input(json.dumps(data))
+        assert "is not one of" in str(err)
+
+        # should fail w/ bmds3
+        data = deepcopy(complete_continuous)
+        data.update(
+            bmds_version=bmds.constants.BMDS312,
+            models={"frequentist_restricted": [bmds.constants.M_Hill]},
+            bmr=c_bmr,
+        )
+        with pytest.raises(ValidationError) as err:
+            validators.validate_input(json.dumps(data))
 
 
-def test_base_validator(complete_continuous):
-    # check validity
-    try:
-        jsonschema.validate(complete_continuous, validators.base_schema)
-    except jsonschema.exceptions.ValidationError:
-        pytest.fail("Should be valid.")
+class TestSessionValidation:
+    def test_base_validator(self, complete_continuous):
+        # check validity
+        assert validators.validate_session(complete_continuous) is None
 
-    # missing required field check
-    data = deepcopy(complete_continuous)
-    data.pop("bmds_version")
-    with pytest.raises(jsonschema.exceptions.ValidationError):
-        jsonschema.validate(data, validators.base_schema)
+        # missing required field check
+        data = deepcopy(complete_continuous)
+        data.pop("bmds_version")
+        with pytest.raises(ValidationError):
+            validators.validate_session(data)
 
 
 class TestDatasetValidation:
     def test_continuous_validator(self, complete_continuous):
-        # check that continuous validators work
+        dataset_type = complete_continuous["dataset_type"]
         datasets = complete_continuous["datasets"]
+
         # check validity
-        try:
-            jsonschema.validate(datasets, validators.continuous_dataset_schema)
-        except jsonschema.exceptions.ValidationError:
-            pytest.fail("Should be valid.")
+        assert validators.validate_datasets(dataset_type, datasets) is None
 
         # missing required field check
         data = deepcopy(datasets)
         data[0].pop("stdevs")
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(data, validators.continuous_dataset_schema)
+        with pytest.raises(ValidationError):
+            validators.validate_datasets(dataset_type, data)
 
-        # n>0 check
+        # n > 0 check
         data = deepcopy(datasets)
         data[0]["ns"][1] = 0
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(data, validators.continuous_dataset_schema)
+        with pytest.raises(ValidationError):
+            validators.validate_datasets(dataset_type, data)
 
     def test_continuous_individual_validator(self, complete_continuous_individual):
-        # check that continuous individual validators work
+        dataset_type = complete_continuous_individual["dataset_type"]
         datasets = complete_continuous_individual["datasets"]
+
         # check validity
-        try:
-            jsonschema.validate(datasets, validators.continuous_individual_dataset_schema)
-        except jsonschema.exceptions.ValidationError:
-            pytest.fail("Should be valid.")
+        assert validators.validate_datasets(dataset_type, datasets) is None
 
         # missing required field check
         data = deepcopy(datasets)
         data[0].pop("responses")
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(data, validators.continuous_individual_dataset_schema)
+        with pytest.raises(ValidationError):
+            validators.validate_datasets(dataset_type, data)
 
         data = deepcopy(datasets)
         data[0].pop("doses")
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(data, validators.continuous_individual_dataset_schema)
+        with pytest.raises(ValidationError):
+            validators.validate_datasets(dataset_type, data)
 
     def test_dichotomous_validator(self, complete_dichotomous):
-        # check that dichotomous validators work
+        dataset_type = complete_dichotomous["dataset_type"]
         datasets = complete_dichotomous["datasets"]
+
         # check validity
-        try:
-            jsonschema.validate(datasets, validators.dichotomous_dataset_schema)
-        except jsonschema.exceptions.ValidationError:
-            pytest.fail("Should be valid.")
+        assert validators.validate_datasets(dataset_type, datasets) is None
 
         # missing required field check
         data = deepcopy(datasets)
         data[0].pop("ns")
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(data, validators.dichotomous_dataset_schema)
+        with pytest.raises(ValidationError):
+            validators.validate_datasets(dataset_type, data)
 
         # n>0 check
         data = deepcopy(datasets)
         data[0]["ns"][1] = 0
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(data, validators.dichotomous_dataset_schema)
+        with pytest.raises(ValidationError):
+            validators.validate_datasets(dataset_type, data)
 
-    def test_dataset_ids(
-        self, complete_continuous, complete_dichotomous, complete_continuous_individual
-    ):
+    def test_dataset_ids(self, complete_dichotomous):
         # Check that commonly used IDs can be used.
-        sets = [
-            (complete_continuous["datasets"], validators.continuous_dataset_schema),
-            (complete_dichotomous["datasets"], validators.dichotomous_dataset_schema),
-            (
-                complete_continuous_individual["datasets"],
-                validators.continuous_individual_dataset_schema,
-            ),
-        ]
-        for datasets, validator in sets:
-            # check missing ID
-            data = deepcopy(datasets)
-            data[0].pop("id")
-            jsonschema.validate(data, validator)
 
-            # check string ID
-            data = deepcopy(datasets)
-            data[0]["id"] = "string"
-            jsonschema.validate(data, validator)
+        dataset_type = complete_dichotomous["dataset_type"]
+        datasets = complete_dichotomous["datasets"]
 
-            # check int ID
-            data = deepcopy(datasets)
-            data[0]["id"] = 123  # int
-            jsonschema.validate(data, validator)
+        # check missing ID
+        data = deepcopy(datasets)
+        data[0].pop("id")
+        validators.validate_datasets(dataset_type, data)
 
-            # check float id
-            data = deepcopy(datasets)
-            data[0]["id"] = 123.1  # float
-            with pytest.raises(jsonschema.exceptions.ValidationError):
-                jsonschema.validate(data, validator)
+        # check string ID
+        data = deepcopy(datasets)
+        data[0]["id"] = "string"
+        validators.validate_datasets(dataset_type, data)
+
+        # check int ID
+        data = deepcopy(datasets)
+        data[0]["id"] = 123  # int
+        validators.validate_datasets(dataset_type, data)
+
+        # check float id
+        data = deepcopy(datasets)
+        data[0]["id"] = 123.1  # float
+        with pytest.raises(ValidationError):
+            validators.validate_datasets(dataset_type, data)
 
 
 class TestModelValidation:
     def test_bmds2(self, complete_continuous):
-        # Check models can be specified
-        cmodels = [{"name": "Exponential-M2"}, {"name": "Exponential-M3"}]
 
-        dmodels = [{"name": "Logistic"}, {"name": "LogLogistic"}]
-
-        # complete check
-        data = deepcopy(complete_continuous)
-        data["models"] = cmodels
-        assert validators.validate_input(json.dumps(data)) is None
-
-        data = deepcopy(complete_continuous)
-        data["models"] = dmodels
-        with pytest.raises(ValidationError):
-            validators.validate_input(json.dumps(data))
+        bmds_version = bmds.constants.BMDS270
+        continuous_type = bmds.constants.CONTINUOUS
+        dichotomous_type = bmds.constants.DICHOTOMOUS
+        c_models = [{"name": "Exponential-M2"}, {"name": "Exponential-M3"}]
+        d_models = [{"name": "Logistic"}, {"name": "LogLogistic"}]
 
         # continuous
-        try:
-            jsonschema.validate(cmodels, validators.c_model_schema)
-        except jsonschema.exceptions.ValidationError:
-            pytest.fail("Should be valid.")
-
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(dmodels, validators.c_model_schema)
+        assert validators.validate_models(bmds_version, continuous_type, c_models) is None
+        with pytest.raises(ValidationError) as err:
+            validators.validate_models(bmds_version, continuous_type, d_models)
+        assert "is not one of" in str(err)
 
         # dichotomous
-        try:
-            jsonschema.validate(dmodels, validators.d_model_schema)
-        except jsonschema.exceptions.ValidationError:
-            pytest.fail("Should be valid.")
-
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(cmodels, validators.d_model_schema)
+        assert validators.validate_models(bmds_version, dichotomous_type, d_models) is None
+        with pytest.raises(ValidationError) as err:
+            validators.validate_models(bmds_version, dichotomous_type, c_models)
+        assert "is not one of" in str(err)
 
     def test_bmds3_dichotomous(self, complete_dichotomous):
         dtype = bmds.constants.DICHOTOMOUS
@@ -170,34 +189,34 @@ class TestModelValidation:
 
         # test success
         assert (
-            validators._validate_models(version, dtype, {"frequentist_restricted": [logprobit]},)
+            validators.validate_models(version, dtype, {"frequentist_restricted": [logprobit]},)
             is None
         )
 
         # assert wrong model type
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version, dtype, {"frequentist_restricted": [bmds.constants.M_Power]},
             )
-        assert "'Power' is not one of" in str(err)
+        assert "is not one of" in str(err)
 
         # assert duplicates model type
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version, dtype, {"frequentist_restricted": [logprobit, logprobit]},
             )
         assert "has non-unique elements" in str(err)
 
         # assert empty
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version, dtype, {"frequentist_restricted": []},
             )
         assert "At least one model must be selected" in str(err)
 
         # assert bayesian duplicates
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version,
                 dtype,
                 {
@@ -212,7 +231,7 @@ class TestModelValidation:
 
         # assert bayesian prior_weight sum
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version,
                 dtype,
                 {
@@ -232,34 +251,33 @@ class TestModelValidation:
 
         # test success
         assert (
-            validators._validate_models(version, dtype, {"frequentist_restricted": [power]},)
-            is None
+            validators.validate_models(version, dtype, {"frequentist_restricted": [power]},) is None
         )
 
         # assert wrong model type
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version, dtype, {"frequentist_restricted": [bmds.constants.M_Probit]},
             )
-        assert "'Probit' is not one of" in str(err)
+        assert "is not one of" in str(err)
 
         # assert duplicates model type
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version, dtype, {"frequentist_restricted": [power, power]},
             )
         assert "has non-unique elements" in str(err)
 
         # assert empty
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version, dtype, {"frequentist_restricted": []},
             )
         assert "At least one model must be selected" in str(err)
 
         # assert bayesian duplicates
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version,
                 dtype,
                 {
@@ -274,7 +292,7 @@ class TestModelValidation:
 
         # assert bayesian prior_weight sum
         with pytest.raises(ValidationError) as err:
-            validators._validate_models(
+            validators.validate_models(
                 version,
                 dtype,
                 {
@@ -288,44 +306,53 @@ class TestModelValidation:
 
 
 class TestBmrOptions:
-    def test_bmds2(self, complete_continuous):
-        # Check models can be specified
-        cbmr = {"type": "Std. Dev.", "value": 1.0}
-
-        dbmr = {"type": "Added", "value": 0.1}
-
-        # complete check
-        data = deepcopy(complete_continuous)
-        data["bmr"] = cbmr
-        assert validators.validate_input(json.dumps(data)) is None
-
-        data = deepcopy(complete_continuous)
-        data["bmr"] = dbmr
-        with pytest.raises(ValidationError):
-            validators.validate_input(json.dumps(data))
+    def test_bmds2(self):
+        c_bmr = {"type": "Std. Dev.", "value": 1.0}
+        d_bmr = {"type": "Added", "value": 0.1}
 
         # continuous
-        try:
-            jsonschema.validate(cbmr, validators.c_bmr_schema)
-        except jsonschema.exceptions.ValidationError:
-            pytest.fail("Should be valid.")
-
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(dbmr, validators.c_bmr_schema)
+        assert validators.validate_bmrs(bmds.constants.CONTINUOUS, c_bmr) is None
+        with pytest.raises(ValidationError) as err:
+            validators.validate_bmrs(bmds.constants.CONTINUOUS, d_bmr)
+        assert "'Added' is not one of" in str(err)
 
         # dichotomous
-        try:
-            jsonschema.validate(dbmr, validators.d_bmr_schema)
-        except jsonschema.exceptions.ValidationError:
-            pytest.fail("Should be valid.")
-
-        with pytest.raises(jsonschema.exceptions.ValidationError):
-            jsonschema.validate(cbmr, validators.d_bmr_schema)
-
-    def test_bmds3(self):
-        # not used in bmds3
-        pass
+        assert validators.validate_bmrs(bmds.constants.DICHOTOMOUS, d_bmr) is None
+        with pytest.raises(ValidationError) as err:
+            validators.validate_bmrs(bmds.constants.DICHOTOMOUS, c_bmr)
+        assert "'Std. Dev.' is not one of" in str(err)
 
 
 class TestOptionSetValidation:
-    pass
+    def test_dichotomous(self):
+        # test success
+        data = [dict(bmr_type="Extra", bmr_value=0.1, confidence_level=0.05, background="Zero")]
+        assert validators.validate_options(bmds.constants.DICHOTOMOUS, data) is None
+
+        # must have at least one option
+        data = []
+        with pytest.raises(ValidationError) as err:
+            validators.validate_options(bmds.constants.DICHOTOMOUS, data)
+        assert "[] is too short" in str(err)
+
+    def test_continuous(self):
+        # test success
+        data = [
+            {
+                "bmr_type": "Std. Dev.",
+                "bmr_value": 1.0,
+                "tail_probability": 0.95,
+                "confidence_level": 0.05,
+                "distribution": "Normal",
+                "variance": "Calculated",
+                "polynomial_restriction": "Use dataset adverse direction",
+                "background": "Estimated",
+            }
+        ]
+        assert validators.validate_options(bmds.constants.CONTINUOUS, data) is None
+
+        # must have at least one option
+        data = []
+        with pytest.raises(ValidationError) as err:
+            validators.validate_options(bmds.constants.CONTINUOUS, data)
+        assert "[] is too short" in str(err)
